Migrate tests to esm
I migrated sources to esm locally and output size with rollup is almost the same
as with browserify.

In this diff I only migrated tests to es modules and run them with the help of
"esm" package. Later when everything will be migrated we can try to use
native node support.