Fix exception, closing of lists in certain edge cases
If a different list item is found, the list parser signals to the
document parser that flow, which might still be open for the previous
item, should close.
This did not account for the fact that other containers, such as lists
or block quotes, might also still be open in the previous item.
This commit solves this by closing other open containers when requested.

This commit also fixes an exception when the signal to close flow was
received, but no flow was open.