Refactor to reuse char codes again
This was introduced in 839564d due to xo.
However, we use a lot of indexing into strings to handle markdown.
Strings in JS use char codes, whether we like it or not, internally.
Using code points messes with indexing, and is very likely to introduce
bugs.

At some point, we should either switch to arrays of code points
(perhaps slow?) or with u8 bytes (same as `markdown-rs`), which
may be nice as we currently support u8 arrays since 910dbf0 anyway.